@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 import {CanActivate} from '@angular/router';
-import {AngularFire} from 'angularfire2';
+import {AngularFireAuth} from 'angularfire2/auth';
 import {Router} from '@angular/router';
 import {Observable, AsyncSubject} from 'rxjs';
 
@@ -8,29 +8,27 @@ import {Observable, AsyncSubject} from 'rxjs';
 export class CanActivateViaAuthGuard implements CanActivate {
   // canActivateVal: boolean = false;
   isLogged: AsyncSubject<any> = new AsyncSubject();
-  afAuth;
 
-  constructor(private af: AngularFire, private router: Router) {
-    this.afAuth = af.auth;
-    this.router;
+  constructor(private afAuth: AngularFireAuth, private router: Router) {
+    // this.afAuth = afAuth.auth;
     this.authenticated();
   }
 
   authenticated() {
-    if(this.afAuth) {
-      this.afAuth.subscribe(
-      auth => {
-        if(!!auth) {
-          this.isLogged.next(true);
-        } else {
+    if (this.afAuth) {
+      this.afAuth.authState.subscribe(
+        auth => {
+          if (!!auth) {
+            this.isLogged.next(true);
+          } else {
+            this.isLogged.next(false);
+          }
+          this.isLogged.complete();
+        },
+        error => {
           this.isLogged.next(false);
+          this.isLogged.complete();
         }
-        this.isLogged.complete();
-      },
-      error => {
-        this.isLogged.next(false);
-        this.isLogged.complete();
-      }
       )
     } else {
       this.isLogged.next(false);
@@ -40,9 +38,11 @@ export class CanActivateViaAuthGuard implements CanActivate {
 
   canActivate(): Observable<boolean> {
     return this.isLogged.do(
-    isLoggedIn => {
-      if(!isLoggedIn) this.router.navigate([''])
-    },
-    error => console.error(error));
+      isLoggedIn => {
+        if (!isLoggedIn) {
+          this.router.navigate(['']);
+        }
+      },
+      error => console.error(error));
   }
 }
