@@ -4,5 +4,13 @@
 // The list of which env maps to which file can be found in `angular-cli.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  firebase: {
+    apiKey: 'AIzaSyClMUeCVSKC1F50BHIH9gjWyoBn1rrpc9o',
+    authDomain: 'cakeorderform2.firebaseapp.com',
+    databaseURL: 'https://cakeorderform2.firebaseio.com',
+    projectId: 'cakeorderform2',
+    storageBucket: 'cakeorderform2.appspot.com',
+    messagingSenderId: '680644753317'
+  }
 };
